Retrieve Standartox toxicity values

Retrieve toxicity values from the Standartox data base http://standartox.uni-landau.de/.

stx_query(
  vers = NULL,
  casnr = NULL,
  concentration_unit = NULL,
  concentration_type = NULL,
  duration = NULL,
  endpoint = c("XX50", "NOEX", "LOEX"),
  effect = NULL,
  exposure = NULL,
  chemical_role = NULL,
  chemical_class = NULL,
  taxa = NULL,
  ecotox_grp = NULL,
  trophic_lvl = NULL,
  habitat = NULL,
  region = NULL,
  ...
)

Arguments

vers
integer; Choose the version of the EPA Ecotox on which Standartox is based on. NULL (default) accesses the most recent version.

concentration_unit
character; Limit data base query to specific concentration units (e.g. ug/l - default).

concentration_type
character; Limit data base query to specific concentration types, can be one of NULL (default), 'active ingredient', 'formulation', 'total', 'not reported', 'unionized', 'dissolved', 'labile'. See https://cfpub.epa.gov/ecotox/pdf/codeappendix.pdf p.4.

duration
integer vector of length two; Limit data base query to specific test durations (hours) (e.g. c(24, 48)). NULL (default).

endpoint
character; Choose endypoint type, must be one of 'XX50' (default), 'NOEX', 'LOEX'.

effect
character; Limit data base query to specific effect groups, multiple entries possible (e.g. 'Mortality', 'Intoxication', 'Growth'). See https://cfpub.epa.gov/ecotox/pdf/codeappendix.pdf p.95. NULL (default).

exposure
character; Choose exposure type, (e.g. aquatic, environmental, diet). NULL (default).

chemical_role
character; Limit data base query to specific chemical roles (e.g. insecticide), multiple entries possible. NULL (default).

chemical_class
character; Limit data base query to specific chemical classes (e.g. neonicotinoid), multiple entries possible. NULL (default).

taxa
character; Limit data base query to specific taxa, multiple entries possible. NULL (default).

ecotox_grp
character; Convenience grouping of organisms in ecotoxicology, must be one of NULL (default), 'invertebrate', 'fish', 'plant_land', 'macrophyte', 'algae'.

habitat
character; Limit data base query to specific organism habitats, can be one of NULL (default) 'marine', 'brackish', 'freshwater', 'terrestrial'.

region
character; Limit data base query to organisms occurring in specific regions, can be one of NULL (default) 'africa', 'america_north', 'america_south', 'asia', 'europe', 'oceania'.

...
currently not used

cas
character, integer; Limit data base query to specific CAS numbers, multiple entries possible (e.g. 1071-83-6, 1071836), NULL (default).

tropic_lvl
character; Trophic level of organism, must be one of NULL (default), 'autotroph', 'heterotroph'.

Value

Returns a list of three data.tables (filtered data base query results, aggregated data base query results, meta information)

Examples

# \donttest{
# might fail if API is not available
stx_query('1071-83-6')

#> Standartox query running..
#> Parameters:
#> vers: 1071-83-6
#> endpoint: XX50

#> Warning: Provided version number not in data.

#> Warning: 400

#> Error in stx_query("1071-83-6"): object 'filtered' not found

stx_query('1071-83-6', duration = c(48, 120))

#> Standartox query running..
#> Parameters:
#> vers: 1071-83-6
#> duration: 48, 120
#> endpoint: XX50

#> Warning: Provided version number not in data.

#> Warning: 400

#> Error in stx_query("1071-83-6", duration = c(48, 120)): object 'filtered' not found

stx_query('1071-83-6', duration = c(48, 120), endpoint = 'XX50')

#> Standartox query running..
#> Parameters:
#> vers: 1071-83-6
#> duration: 48, 120
#> endpoint: XX50

#> Warning: Provided version number not in data.

#> Warning: 400

#> Error in stx_query("1071-83-6", duration = c(48, 120), endpoint = "XX50"): object 'filtered' not found

# }